Split outlets to separate services
