+++ src/accessories/delta2maxAccessory.ts
@@ -4,9 +4,11 @@ import { EcoFlowHomebridgePlatform } from '../platform.js';
 import { EcoFlowAccessory } from './ecoFlowAccessory.js';
 import { DeviceConfig } from '../config.js';
 import { BatteryService } from './services/batteryService.js';
-import { OutletsService } from './services/outletsService.js';
 import { ServiceBase } from './services/serviceBase.js';
 import { EcoFlowMqttApi } from './apis/ecoFlowMqttApi.js';
+import { OutletUsbService } from './services/outletUsbService.js';
+import { OutletCarService } from './services/outletCarService.js';
+import { OutletAcService } from './services/outletAcService.js';
 
 export class Delta2MaxAccessory extends EcoFlowAccessory {
   protected override registerServices(
@@ -18,7 +20,9 @@ export class Delta2MaxAccessory extends EcoFlowAccessory {
     const result = [];
 
     result.push(new BatteryService(accessory, platform, config, api));
-    result.push(new OutletsService(accessory, platform, config, api));
+    result.push(new OutletUsbService(accessory, platform, config, api));
+    result.push(new OutletAcService(accessory, platform, config, api));
+    result.push(new OutletCarService(accessory, platform, config, api));
 
     return result;
   }
